refactor: rename to md consistent with the other compilers using extname
